+++ packages/marshal/test/test-marshal.js
@@ -9,7 +9,7 @@ import {
   passStyleOf,
 } from '../src/marshal';
 
-const { is, isFrozen, create, prototype: objectPrototype } = Object;
+const { is, freeze, isFrozen, create, prototype: objectPrototype } = Object;
 
 // this only includes the tests that do not use liveSlots
 
@@ -154,6 +154,29 @@ test('serialize static data', t => {
   // well known unsupported symbols
   t.throws(() => ser(Symbol.iterator), { message: /Unsupported symbol/ });
 
+  const cd = ser(harden([1, 2]));
+  t.is(isFrozen(cd), true);
+  t.is(isFrozen(cd.slots), true);
+});
+
+test('unserialize static data', t => {
+  const m = makeMarshal();
+  const uns = body => m.unserialize({ body, slots: [] });
+
+  // should be frozen
+  const arr = uns('[1,2]');
+  t.truthy(isFrozen(arr));
+  const a = uns('{"b":{"c":{"d": []}}}');
+  t.truthy(isFrozen(a));
+  t.truthy(isFrozen(a.b));
+  t.truthy(isFrozen(a.b.c));
+  t.truthy(isFrozen(a.b.c.d));
+});
+
+test('serialize errors', t => {
+  const m = makeMarshal();
+  const ser = val => m.serialize(val);
+
   t.deepEqual(ser(harden(Error())), {
     body:
       '{"@qclass":"error","errorId":"error:anon-marshal#1","message":"","name":"Error"}',
@@ -166,12 +189,33 @@ test('serialize static data', t => {
     slots: [],
   });
 
-  const cd = ser(harden([1, 2]));
-  t.is(isFrozen(cd), true);
-  t.is(isFrozen(cd.slots), true);
+  // TODO Once https://github.com/Agoric/SES-shim/issues/579 is merged
+  // do a golden of the error notes generated by the following
+
+  // Extra properties
+  const errExtra = Error('has extra properties');
+  errExtra.foo = [];
+  freeze(errExtra);
+  t.assert(isFrozen(errExtra));
+  t.falsy(isFrozen(errExtra.foo));
+  t.deepEqual(ser(errExtra), {
+    body:
+      '{"@qclass":"error","errorId":"error:anon-marshal#3","message":"has extra properties","name":"Error"}',
+    slots: [],
+  });
+  t.falsy(isFrozen(errExtra.foo));
+
+  // Bad prototype and bad "message" property
+  const nonErrorProto1 = { __proto__: Error.prototype, name: 'included' };
+  const nonError1 = { __proto__: nonErrorProto1, message: [] };
+  t.deepEqual(ser(harden(nonError1)), {
+    body:
+      '{"@qclass":"error","errorId":"error:anon-marshal#4","message":"","name":"included"}',
+    slots: [],
+  });
 });
 
-test('unserialize static data', t => {
+test('unserialize errors', t => {
   const m = makeMarshal();
   const uns = body => m.unserialize({ body, slots: [] });
 
@@ -189,15 +233,6 @@ test('unserialize static data', t => {
   const em3 = uns('{"@qclass":"error","message":"msg3","name":"Unknown"}');
   t.truthy(em3 instanceof Error);
   t.is(em3.message, 'msg3');
-
-  // should be frozen
-  const arr = uns('[1,2]');
-  t.truthy(isFrozen(arr));
-  const a = uns('{"b":{"c":{"d": []}}}');
-  t.truthy(isFrozen(a));
-  t.truthy(isFrozen(a.b));
-  t.truthy(isFrozen(a.b.c));
-  t.truthy(isFrozen(a.b.c.d));
 });
 
 test('serialize ibid cycle', t => {
